Move the existing unit tests into tests.unit
