+++ tests/unit/accounts/test_models.py
@@ -14,7 +14,7 @@ import pytest
 
 from warehouse.accounts.models import User, UserFactory
 
-from ..common.db.accounts import (
+from ...common.db.accounts import (
     UserFactory as DBUserFactory, EmailFactory as DBEmailFactory,
 )
