+++ tests/unit/cache/origin/test_fastly.py
