+++ tests/unit/cli/test_db.py
