+++ tests/unit/cli/test_shell.py
