+++ tests/unit/legacy/api/test_simple.py
@@ -16,8 +16,8 @@ from pyramid.httpexceptions import HTTPMovedPermanently
 
 from warehouse.legacy.api import simple
 
-from ...common.db.accounts import UserFactory
-from ...common.db.packaging import (
+from ....common.db.accounts import UserFactory
+from ....common.db.packaging import (
     ProjectFactory, ReleaseFactory, FileFactory, JournalEntryFactory,
 )
