+++ tests/unit/utils/test_gravatar.py
