+++ tests/unit/utils/test_static.py
